format use proposals and cleanup
